Feature: adds Contributing.md & improves README.md
This commit brings the following changes:

- Creates a CONTRIBUTING.md to provide certain guidlines for the
procedures that should be taken before submitting a pull request,
issue and generally socially interacting with the project's
community. This is often a good idea to have in any open source
project. For more info, read about CONTRIBUTING.md here:
https://docs.github.com/en/github/building-a-strong-community/setting-guidelines-for-repository-contributors

- Modifies the README.md file to be more expressive and appealing
by adding the logo image, some badges to provide general info about
the project. This also modifies the previously known as "Using npm"
to a more general "Installation" category that provides more information
about the possible ways this package can be used. This also adds a Contributing
section that will redirect users to the CONTRIBUTING.md file easily.
Changes to be committed:
	new file:   CONTRIBUTING.md
	modified:   README.md